+++ src/Oro/Bundle/LayoutBundle/Layout/Extension/SvgIconsSupportContextConfigurator.php
@@ -7,6 +7,7 @@ namespace Oro\Bundle\LayoutBundle\Layout\Extension;
 use Oro\Bundle\LayoutBundle\Provider\SvgIconsSupportProvider;
 use Oro\Component\Layout\ContextConfiguratorInterface;
 use Oro\Component\Layout\ContextInterface;
+use Oro\Component\Layout\Extension\Theme\Model\ThemeManager;
 use Symfony\Component\OptionsResolver\Options;
 
 /**
@@ -15,8 +16,15 @@ use Symfony\Component\OptionsResolver\Options;
  */
 class SvgIconsSupportContextConfigurator implements ContextConfiguratorInterface
 {
+    private ThemeManager $themeManager;
+
     public function __construct(private SvgIconsSupportProvider $svgIconsSupportProvider)
     {
+    }
+
+    public function setThemeManager(ThemeManager $themeManager): void
+    {
+        $this->themeManager = $themeManager;
     }
 
     #[\Override]
@@ -40,7 +48,7 @@ class SvgIconsSupportContextConfigurator implements ContextConfiguratorInterface
                             return false;
                         }
 
-                        return $this->svgIconsSupportProvider->isSvgIconsSupported($themeName);
+                        return $this->themeManager->getThemeOption($themeName, 'svg_icons_support') ?? false;
                     },
                 ]
             )