+++ src/Oro/Component/Layout/Extension/Theme/Model/ThemeFactory.php
@@ -51,6 +51,7 @@ class ThemeFactory implements ThemeFactoryInterface
             'directory',
             'groups',
             'description',
+            'fonts'
         ];
 
         foreach ($properties as $property) {